+++ torchbearer/callbacks/early_stopping.py
@@ -29,9 +29,6 @@ class EarlyStopping(Callback):
         self.verbose = verbose
         self.mode = mode
 
-        self.wait = 0
-        self.stopped_epoch = 0
-
         if self.mode not in ['min', 'max']:
             if 'acc' in self.monitor:
                 self.mode = 'max'
@@ -45,10 +42,22 @@ class EarlyStopping(Callback):
             self.min_delta *= 1
             self.monitor_op = lambda x1, x2: x1 > x2
 
-    def on_start(self, state):
         self.wait = 0
         self.stopped_epoch = 0
         self.best = float('inf') if self.mode == 'min' else -float('inf')
+
+    def state_dict(self):
+        state_dict = {
+            'wait': self.wait,
+            'best': self.best,
+            'stopped_epoch': self.stopped_epoch
+        }
+        return state_dict
+
+    def load_state_dict(self, state_dict):
+        self.wait = state_dict['wait']
+        self.best = state_dict['best']
+        self.stopped_epoch = state_dict['stopped_epoch']
 
     def on_end_epoch(self, state):
         current = state[torchbearer.METRICS][self.monitor]